Removed harp-static and add express for router
Now we can handle posts to /contact, etc. In theory we can reuse the views and make them dynamic, but I like the idea of keeping all the content in the static form, and only using routes for posting data.

The only question would be the 404 page, but I think that could be done withe an XHR request and shave off a random 5 for the sidebar (or something).

See routes.js for custom route handling.